restructure project for deno-node support
