+++ .gitignore
@@ -1,4 +1,4 @@
 node_modules
 coverage
-lib
 dist
+lib
