+++ .npmignore
@@ -1,3 +1,10 @@
+.github
+.vscode
 coverage
 tests
 src
+
+mod.ts
+
+Makefile
+tsconfig.json
