--- src/constants.ts
@@ -1,6 +1,0 @@
-/**
- * Default byte length of each ArrayBuffer.
- *
- * @type {number}
- */
-export const BYTE_LENGTH: number = 1048576
