--- src/index.ts
@@ -1,4 +1,0 @@
-export * from './constants'
-export * from './Reader'
-export * from './Writer'
-export * from './DynWriter'
