+++ src/reader.ts
